The IRP structures

Commodity

InventoryRoutingLNS.Commodity — Type

Commodity

A commodity (or product) is released by depots, demanded by customers and transported by routes.

We consider the multi-commodity IRP in this package.

Fields

m::Int: index of the commodity.

l::Int: length of a commodity.

Base.copy — Method

Base.copy(commodity::Commodity)

Copy commodity.

Base.show — Method

Base.show(io::IO, commodity::Commodity)

Display commodity in the terminal.

Site

InventoryRoutingLNS.Site — Type

Site

A site is a generic place where inventory is monitored.

InventoryRoutingLNS.positive_inventory — Method

positive_inventory(site::Site)

Check if the site inventory is nonnegative.

Depot

InventoryRoutingLNS.Depot — Type

Depot

A depot releases commodities every day.

In the IRP we plan route deliveries starting at depots in a centralized manner.

Fields

d::Int: index of the depot.

v::Int: index of the site (among depots and customers).

coordinates::Tuple{Float64,Float64}: coordinates of the depot.

excess_inventory_cost::Vector{Int}: unit excess inventory cost per commodity.

initial_inventory::Vector{Int}: initial inventory per commodity.

production::Matrix{Int}: release per commodity and day.

maximum_inventory::Matrix{Int}: maximum inventory per commodity and day.

quantity_sent::Matrix{Int}: quantity sent (with current solution) per commodity and day.

inventory::Matrix{Int}: inventory (with current solution) per commodity and day.

commodity_used::Vector{Bool}: encode the commodities used by the depot.

Base.copy — Method

Base.copy(depot::Depot)

Copy depot.

Base.show — Method

Base.show(io::IO, depot::Depot)

Display depot in the terminal.

InventoryRoutingLNS.renumber — Method

renumber(depot::Depot; d::Int, v::Int)

Set d and v as indices for depot over depots and sites respectively.

This can be used when dividing an instance into several smaller ones.

Customer

InventoryRoutingLNS.Customer — Type

Customer

A customer demands commodities every day.

In the IRP we plan route deliveries to customers in a centralized manner.

Fields

c::Int: index of the customer.

v::Int: index of the site (among depots and customers).

coordinates::Tuple{Float64,Float64}: coordinates of the customer.

excess_inventory_cost::Vector{Int}: unit excess inventory cost per commodity.

shortage_cost::Vector{Int}: unit shortage cost per commodity.

initial_inventory::Vector{Int}: initial inventory per commodity.

demand::Matrix{Int}: demand per commodity and day.

maximum_inventory::Matrix{Int}: maximum inventory per commodity and day.

quantity_received::Matrix{Int}: quantity received (with current solution) per commodity and day.

inventory::Matrix{Int}: inventory (with current solution) per commodity and day.

commodity_used::Vector{Bool}: encode the commodities used by the customer.

Base.copy — Method

Base.copy(customer::Customer)

Copy customer.

Base.show — Method

Base.show(io::IO, customer::Customer)

Display customer in the terminal.

InventoryRoutingLNS.positive_inventory_zero_demand_and_initial_inventory — Method

positive_inventory_zero_demand_and_initial_inventory(customer::Customer, m::Int)

Check if customer has received some commodity m whereas it does not need it.

It is used for feasibility tests in feasibility.

InventoryRoutingLNS.positive_inventory_zero_demand_and_initial_inventory — Method

positive_inventory_zero_demand_and_initial_inventory(customer::Customer)

Check if customer has received one type of commodity it does not need.

It is used for feasibility tests in feasibility.

InventoryRoutingLNS.renumber — Method

renumber(customer::Customer; c::Int, v::Int)

Set c and v as indices for customer over customers and sites respectively.

This can be used when dividing an instance into several smaller ones.

Route

InventoryRoutingLNS.Route — Type

Route

A route corresponds to a path and deliveries per customer.

Fields

id::Int: id of the route.

t::Int: start date.

d::Int: start depot.

stops::Vector{RouteStop}: stops processed by the route.

InventoryRoutingLNS.RouteStop — Type

RouteStop

Stop where a delivery is processed.

Fields

c::Int: customer where the stop is made.

t::Int: arrival day in the stop.

Q::Vector{Int}: quantities to deliver per commodity.

Base.copy — Method

Base.copy(stop::RouteStop)

Copy route stop.

Base.copy — Method

Base.copy(route::Route)

Copy route.

Base.show — Method

Base.show(io::IO, route::Route)

Display route in the terminal.

InventoryRoutingLNS.check_if_connected — Method

check_if_connected(routes1::Vector{Route}, routes2::Vector{Route})

Check if two sets of routes are connected by the sites they involve.

Two sets of routes are connected if they share any site (depot or customer).

InventoryRoutingLNS.decouple_routes — Method

decouple_routes(routes::Vector{Route})

Decouple a set of routes into sets of routes that do not share any depot or customer.

It can be used to decompose a MILP into smaller ones. For instance, see refill_routes!.

InventoryRoutingLNS.get_M — Method

get_M(stop::RouteStop)

Get the number of commodities M from stop.

InventoryRoutingLNS.get_nb_stops — Method

get_nb_stops(route::Route)

Get the number of (customer) stops of route.

InventoryRoutingLNS.get_nb_unique_stops — Method

get_nb_unique_stops(route::Route)

Get the number of distinct (customer) stops of route.

InventoryRoutingLNS.mycopy — Method

mycopy(stop::RouteStop)

Custom copy of a route stop.

This is currently used in routing neighborhoods.

InventoryRoutingLNS.mycopy — Method

mycopy(route::Route)

Custom copy of route.

This is currently used in routing neighborhoods.

InventoryRoutingLNS.mycopy — Method

mycopy(stops::Vector{RouteStop})

Custom copy of a vector of route stops.

This is currently used in routing neighborhoods.

InventoryRoutingLNS.unique_stops — Method

unique_stops(route::Route)

Get the set of customer indices visited by route.

InventoryRoutingLNS.unique_stops — Method

unique_stops(routes::Vector{Route})

Get the set of customer indices visited by the set of routes.

InventoryRoutingLNS.uses_commodity — Method

uses_commodity(route::Route, m::Int)

Check if commodity m is delivered to any stop visited by route.

InventoryRoutingLNS.uses_commodity — Method

uses_commodity(stop::RouteStop, m::Int)

Check if commodity m is delivered to stop.

Solution

InventoryRoutingLNS.Solution — Type

Solution

Abstract IRP solution.

A solution gathers routes in different ways. See SimpleSolution or StructuredSolution.

Base.show — Method

Base.show(io::IO, solution::Solution)

Display solution.

Simple Solution

InventoryRoutingLNS.SimpleSolution — Type

SimpleSolution

IRP solution as vector of routes.

Fields

routes::Vector{Route}: routes of the solution.

InventoryRoutingLNS.add_route! — Method

add_route!(solution::SimpleSolution, route::Route)

Add route to a simple solution.

InventoryRoutingLNS.delete_route! — Method

delete_route!(solution::SimpleSolution, r::Int)

Delete the route at index r from a simple solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::SimpleSolution, t::Int, d::Int)

Get the list of routes starting on day t from depot d in a simple solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::SimpleSolution, t::Int)

Get the list of routes starting on day t in a simple solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::SimpleSolution)

Get the list of routes in a simple solution.

InventoryRoutingLNS.nb_routes — Method

nb_routes(solution::SimpleSolution)

Get the number of routes in a simple solution.

Structured Solution

InventoryRoutingLNS.StructuredSolution — Type

StructuredSolution

IRP solution as matrix of routes per day and depot.

Fields

routes_per_day_and_depot::Matrix{Vector{Route}}: routes of the solution.

Base.copy — Method

Base.copy(solution::StructuredSolution)

Copy structured solution.

InventoryRoutingLNS.add_route! — Method

add_route!(solution::StructuredSolution, route::Route)

Push route to a structured solution.

InventoryRoutingLNS.add_route! — Method

add_route!(solution::StructuredSolution, routes::Vector{Route})

Push a set routes to a structured solution.

InventoryRoutingLNS.delete_route! — Method

delete_route!(solution::StructuredSolution, t::Int, d::Int, r::Int)

Delete the r-th route starting on day t by depot d from a structured solution.

InventoryRoutingLNS.delete_route! — Method

delete_route!(solution::StructuredSolution, route::Route)

Delete route from a structured solution.

InventoryRoutingLNS.delete_routes! — Method

delete_routes!(solution::StructuredSolution, t::Int, d::Int, rs)

Delete the rs-th routes starting on day t by depot d from a structured solution.

InventoryRoutingLNS.delete_routes! — Method

delete_routes!(solution::StructuredSolution, routes::Vector{Route})

Delete a set routes from a structured solution.

InventoryRoutingLNS.get_D — Method

get_D(solution::StructuredSolution)

Get the number of depots D from the structured solution.

InventoryRoutingLNS.get_T — Method

get_T(solution::StructuredSolution)

Get the horizon T from the structured solution.

InventoryRoutingLNS.get_route — Method

get_route(solution::StructuredSolution, t::Int, d::Int, r::Int)

Get the r-th route starting on day t by depot d in a structured solution.

InventoryRoutingLNS.get_route_day — Method

get_route_day(solution::StructuredSolution, t::Int, r::Int)

Get the r-th route starting on day t in a structured solution.

InventoryRoutingLNS.get_route_depot — Method

get_route_depot(solution::StructuredSolution, d::Int, r::Int)

Get the r-th route starting by depot d in a structured solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::StructuredSolution, t::Int, d::Int)

Get the list of routes starting on day t from depot d in a structured solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::StructuredSolution, t::Int)

Get the list of routes starting on day t in a structured solution.

InventoryRoutingLNS.list_routes — Method

list_routes(solution::StructuredSolution)

Get the list of routes of a structured solution.

InventoryRoutingLNS.list_routes_customer — Method

list_routes_customer(solution::StructuredSolution, c::Int, t::Int)

Get the list of routes visiting customer c and starting on day t in a structured solution.

InventoryRoutingLNS.list_routes_customer — Method

list_routes_customer(solution::StructuredSolution, c::Int)

Get the list of routes visiting customer c in a structured solution.

InventoryRoutingLNS.list_routes_depot — Method

list_routes_depot(solution::StructuredSolution, d::Int)

Get the list of routes starting by depot d in a structured solution.

InventoryRoutingLNS.mycopy — Method

mycopy(solution::StructuredSolution)

Custom copy of a structured solution.

InventoryRoutingLNS.nb_routes — Method

nb_routes(solution::StructuredSolution, t::Int, d::Int)

Compute the number of routes starting on day t from depot d in a structured solution.

InventoryRoutingLNS.nb_routes — Method

nb_routes(solution::StructuredSolution, t::Int)

Compute the number of routes starting on day t in a structured solution.

InventoryRoutingLNS.nb_routes — Method

nb_routes(solution::StructuredSolution)

Compute the number of routes in a structured solution.

InventoryRoutingLNS.nb_routes_customer — Method

nb_routes_customer(solution::StructuredSolution, t::Int, c::Int)

Compute the number of routes that visit customer c and start on day t in a structured solution.

InventoryRoutingLNS.nb_routes_depot — Method

nb_routes_depot(solution::StructuredSolution, d::Int)

Compute the number of routes starting from depot d in a structured solution.

Instance

InventoryRoutingLNS.Instance — Type

Instance

An IRP instance gathers all information needed to define the problem.

Commodities, depots, customers, time and transport information is stored here. A solution can also be saved.

Fields

T::Int: horizon.

D::Int: number of depots.

C::Int: number of customers.

M::Int: number of commodities.

vehicle_capacity::Int: capacity of a vehicle (1D).

km_cost::Int: unit kilometer cost.

vehicle_cost::Int: cost of making a stop for a delivery.

stop_cost::Int: cost for using a vehicle.

nb_transport_hours_per_day::Int: number of transport hours possible in a day.

S_max::Int: maximum number of stops in a route.

commodities::Vector{Commodity}: the commodities.

depots::Vector{Depot}: the depots.

customers::Vector{Customer}: the customers.

dist::Matrix{Int}: distances between sites in kilometers.

transport_durations::Matrix{Int}: durations between sites in hours.

solution::StructuredSolution: current solution of the instance.

Base.copy — Method

Base.copy(instance::Instance)

Copy instance.

Base.show — Method

Base.show(io::IO, instance::Instance)

Display instance in the terminal.

InventoryRoutingLNS.commodities_used_customer — Method

commodities_used_customer(instance::Instance, c::Int)

Select the commodities with positive initial inventory or positive cumulated demand over the horizon at customer c.

It is used in the customer reinsertion large neighborhood.

InventoryRoutingLNS.commodities_used_depot — Method

commodities_used_depot(instance::Instance, d::Int)

Select the commodities with positive initial inventory or positive cumulated release over the horizon at depot d.

InventoryRoutingLNS.compute_nb_km — Method

compute_nb_km(route::Route, instance::Instance)

Compute the number of kilometers travelled along route.

InventoryRoutingLNS.compute_route_duration — Method

compute_route_duration(route::Route, instance::Instance)

Compute the total duration of route in hours.

InventoryRoutingLNS.content_size — Method

content_size(route::Route, instance::Instance)

Compute the space used in the 1D vehicle of route by all stops deliveries.

InventoryRoutingLNS.content_size — Method

content_size(stop::RouteStop, instance::Instance)

Compute the 1D space occupied by the quantities to be sent to stop.

InventoryRoutingLNS.get_vehicle_cost — Method

get_vehicle_cost(instance::Instance)

Get the vehicle cost data of instance.

It is used in the cost functions as compute_route_cost.

InventoryRoutingLNS.ratio_demanded_released — Method

ratio_demanded_released(instance::Instance)

Estimate the ratio quantity released/quantity demanded per commodity.

We use it as ratio to rescale unbalanced instances in rescale_release_demand! to test algorithms.

InventoryRoutingLNS.rescale_release_demand! — Method

rescale_release_demand!(instance::Instance; verbose::Bool = false)

Rescale instance demand and release.

The aim is to have total demand close to total release. Instances with this property are more difficult to solve.

InventoryRoutingLNS.routes_to_sent_quantities — Method

routes_to_sent_quantities(solution::SimpleSolution, instance::Instance)

Deduce the quantities sent from each depot to each customer per commodity and per day from a simple solution (set of routes).

InventoryRoutingLNS.routes_to_sent_quantities — Method

routes_to_sent_quantities(solution::StructuredSolution, instance::Instance)

Deduce the quantities sent from each depot to each customer per commodity and per day from solution (set of routes).

InventoryRoutingLNS.select_relevant_customers — Method

select_relevant_customers(instance::Instance, m::Int)

Select the customers having positive initial inventory or positive cumulated demand over the horizon in commodity m.

It is used in the graph creation in commodity_flow_graph to sparsify.

InventoryRoutingLNS.select_relevant_depots — Method

select_relevant_depots(instance::Instance, m::Int)

Select the depots having positive initial inventory or positive cumulated release over the horizon in commodity m.

It is used in the graph creation in commodity_flow_graph to sparsify.

Analysis

InventoryRoutingLNS.AverageContentSizes — Type

AverageContentSizes

Gather the average content size in the routes of a solution per depot, day and customer.

Designed to be passed as argument in the second pass of the greedy heuristic modified_capa_initialization_plus_ls!.

Fields

avg_l_d::Vector{Float64}: average content size per depot.

avg_l_c::Vector{Float64}: average content size per customer.

avg_l_t::Vector{Float64}: average content size per day.

InventoryRoutingLNS.compute_average_content_size — Method

compute_average_content_size(instance::Instance)

Compute the average content size over the routes in the solution of instance.

This information can be used as a metric to estimate a solution quality.

InventoryRoutingLNS.compute_average_content_size_customer — Method

compute_average_content_size_customer(instance::Instance)

Compute the average content size over the routes that visit each depot in the solution of instance.

Return a vector with one value per customer. This information is used for the second path of the modified_capa_initialization_plus_ls! to estimate the flow arcs' cost.

InventoryRoutingLNS.compute_average_content_size_day — Method

compute_average_content_size_day(instance::Instance)

Compute the average content size over the routes that start on each day in the solution of instance.

Return a vector with one value per day. This information is used for the second path of the modified_capa_initialization_plus_ls! to estimate the flow arcs' cost.

InventoryRoutingLNS.compute_average_content_size_depot — Method

compute_average_content_size_depot(instance::Instance)

Compute the average content size over the routes that start from each depot in the solution of instance.

Return a vector with one value per depot. This information is used for the second path of the modified_capa_initialization_plus_ls! to estimate the flow arcs' cost.

InventoryRoutingLNS.compute_average_nb_km — Method

compute_average_nb_km(instance::Instance)

Compute the average number of kilometers of the routes in the solution of instance.

InventoryRoutingLNS.compute_average_route_duration — Method

compute_average_route_duration(instance::Instance, solution::SimpleSolution)

Compute the average duration in hours of the routes in solution for instance.

InventoryRoutingLNS.compute_average_route_duration — Method

compute_average_route_duration(instance::Instance)

Compute the average duration in hours of the routes in the solution of instance.

InventoryRoutingLNS.compute_avg_nb_stops — Method

compute_avg_nb_stops(instance::Instance)

Compute the average number of stops over the routes in the solution of instance.

InventoryRoutingLNS.compute_route_durations — Method

compute_route_durations(instance::Instance, solution::SimpleSolution)

Compute the vector of route durations, one element per route of solution.

InventoryRoutingLNS.get_most_expensive_customers — Method

get_most_expensive_customers(instance::Instance)

Get the indices of the customers sorted by their costs in the solution.

The cost of a customer is computed as the sum of the inventory compute_inventory_cost and shortage compute_shortage_cost costs.