Cost, feasibility and inventory dynamics

Cost

InventoryRoutingLNS.compute_cost — Method

compute_cost(instance::Instance;
                ds::Vector{Int} = collect(1:instance.D),
                cs::Vector{Int} = collect(1:instance.C),
                ms::Vector{Int} = collect(1:instance.M),
                ts::Vector{Int} = collect(1:instance.T),
                solution::Solution = instance.solution,
                verbose::Bool = false,
)

Compute the total cost of the solution of instance, choose to print detailed results or not.

Two possible cases:

verbose is true: we call compute_detailed_cost.

verbose is false: we call compute_undetailed_cost.

InventoryRoutingLNS.compute_detailed_cost — Method

compute_detailed_cost(instance::Instance;
                        ds::Vector{Int} = collect(1:instance.D),
                        cs::Vector{Int} = collect(1:instance.C),
                        ms::Vector{Int} = collect(1:instance.M),
                        ts::Vector{Int} = collect(1:instance.T),
                        solution::Solution = instance.solution,
)

Compute the total cost of the solution of instance and print detailed results.

Same principle as for compute_undetailed_cost but with statistics on the decomposition of the cost. It may be useful for tests and investigation.

InventoryRoutingLNS.compute_inventory_cost — Function

compute_inventory_cost(customer::Customer, 
                        ms::Vector{Int} = collect(1:get_M(customer)), 
                        ts::Vector{Int} = collect(1:get_T(customer))
)

Compute the excess inventory cost of customer on the period ts for the commodities indexed by ms.

We sum the inventory costs of the commodities indexed by ms on the days ts.

InventoryRoutingLNS.compute_inventory_cost — Function

compute_inventory_cost(depot::Depot, 
                        ms::Vector{Int} = collect(1:get_M(depot)), 
                        ts::Vector{Int} = collect(1:get_T(depot))
)

Compute the excess inventory cost of depot on the period ts for the commodities indexed by ms.

We sum the inventory costs of the commodities indexed by ms on the days ts.

InventoryRoutingLNS.compute_inventory_cost — Function

compute_inventory_cost(depot::Depot, t::Int, ms::Vector{Int} = collect(1:get_M(depot)))

Compute the excess inventory cost of depot on day t for the commodities indexed by ms.

The excess inventory cost on day t for commodity m is defined as the product of the number of commodities of type m above the maximum inventory and the unit excess cost. The unit excess cost is defined for each depot on each day for each commodity.

InventoryRoutingLNS.compute_inventory_cost — Function

compute_inventory_cost(customer::Customer, t::Int, ms::Vector{Int} = collect(1:get_M(customer)))

Compute the excess inventory cost of customer on day t for the commodities indexed by ms.

The excess inventory cost on day t for commodity m is defined as the product of the number of commodities of type m above the maximum inventory and the unit excess cost. The unit excess cost is defined for each customer on each day for each commodity.

InventoryRoutingLNS.compute_km_cost — Method

compute_km_cost(route::Route, instance::Instance)

Compute the kilometers' cost of route.

The kilometers' cost is defined as the product of the number of km travelled and the unit km cost of instance.

InventoryRoutingLNS.compute_route_cost — Method

compute_route_cost(route::Route, instance::Instance)

Compute the routing cost of route.

The routing cost is defined as the sum of the kilometers' cost, the stops' cost and the vehicle cost.

InventoryRoutingLNS.compute_shortage_cost — Function

compute_shortage_cost(customer::Customer, 
                        ms::Vector{Int} = collect(1:get_M(customer)), 
                        ts::Vector{Int} = collect(1:get_T(customer))
)

Compute the shortage cost of customer on the period ts for the commodities indexed by ms.

We sum the shortage costs of the commodities indexed by ms on the days ts.

InventoryRoutingLNS.compute_shortage_cost — Function

compute_shortage_cost(customer::Customer, t::Int, ms::Vector{Int} = collect(1:get_M(customer)))

Compute the shortage cost of customer on day t for the commodities indexed by ms.

The shortage cost on day t for commodity m is defined as the product of the number of commodities of type m below the minimum inventory on the morning of day t (thus inventory of the previous evening less the demand on day t) and the unit shortage cost. The unit shortage cost is defined for each customer for each commodity.

InventoryRoutingLNS.compute_stops_cost — Method

compute_stops_cost(route::Route, instance::Instance)

Compute the stops' cost of route.

The stops' cost is defined as the product of the number of stops and the unit cost of a stop of instance .

InventoryRoutingLNS.compute_undetailed_cost — Method

compute_undetailed_cost(instance::Instance;
                        ds::Vector{Int} = collect(1:instance.D),
                        cs::Vector{Int} = collect(1:instance.C),
                        ms::Vector{Int} = collect(1:instance.M),
                        ts::Vector{Int} = collect(1:instance.T),
                        solution::Solution = instance.solution,
)

Compute the total cost of the solution of instance.

The total cost is defined as the sum of the routing costs, excess inventory costs at the depots and customers, and shortage costs at the customers. Notably for neighborhoods, when computing the cost, we may want to restrict the instance to:

the depots indexed by ds.

the customers indexed by cs.

the commodities indexed by ms.

the days ts.

Feasibility

InventoryRoutingLNS.delays_coherent — Method

delays_coherent(route::Route, instance::Instance; verbose::Bool = false)

Check if the dates of the deliveries of route are coherent with transport durations.

In the Instance structure, we have a maximum number of transport hours per day nb_transport_hours_per_day, as well as the transport duration for each site-to-site travel in transport_durations. We can therefore follow the path of route and check if the dates and transport durations are coherent.

InventoryRoutingLNS.feasibility — Method

feasibility(customer::Customer; verbose::Bool = false)

Check if the current solution of instance is feasible with respect to customer inventory.

Two conditions have to be verified:

the inventory is nonnegative for each commodity on each day.

the commodities received are restricted to those demanded.

Remark: the second condition is necessary to avoid sending commdodities to customers that do not need them to lower the excess inventory costs at the depots.

InventoryRoutingLNS.feasibility — Method

feasibility(depot::Depot; verbose::Bool = false)

Check if the current solution of instance is feasible with respect to depot inventory.

Before applying this check, the instance and stored solution must be coherent. This can be made manually, or using update_instance_from_solution!. We then check if the depot inventory is nonnegative on each day for each commodity.

InventoryRoutingLNS.feasibility — Method

feasibility(instance::Instance;
            ds::Vector{Int} = collect(1:instance.D),
            cs::Vector{Int} = collect(1:instance.C),
            solution::Solution = instance.solution,
            verbose::Bool = false,
)

Check if solution is feasible for instance.

For this function to make sense, we need a coherence between the inventories of the sites in instance and solution. It can be applied to the whole instance updated with update_instance_from_solution!. It can also be applied locally to a sub-solution corresponding to some routes and sub-sets of depots ds and customers cs, with inventories updated using update_instance_some_routes!.

InventoryRoutingLNS.feasibility — Method

feasibility(route::Route, instance::Instance; verbose::Bool = false)

Check if route is feasible for instance.

A route is feasible if each of the following conditions is respected:

the number of stops does not exceed S_max.

the total content size of the deliveries is below the vehicle capacity.

the arrival date to the last stop is before the horizon T.

the dates of arrival in the stops are coherent with transport durations.

InventoryRoutingLNS.stop_depot_not_compatible — Method

stop_depot_not_compatible(depot::Depot, stop::RouteStop; verbose::Bool = false)

Check if stop contains commodities not used by depot (no release or initial inventory).

This check is implicitly done by feasibility but faster to compute, we therefore do not include it in the feasibility function of a depot, but instead call it when the whole feasibility check is not necessary, see insert_multi_depot! for instance.

Inventory dynamics

InventoryRoutingLNS.coherence_solution_inventory — Method

coherence_solution_inventory(instance::Instance)

Check the coherence between the fields of an instance and its solution.

InventoryRoutingLNS.reset_solution! — Method

reset_solution!(instance::Instance)

Replace the solution of instance with an empty one and update inventory.

InventoryRoutingLNS.solved_instance — Method

solved_instance(instance::Instance, solution::Solution)

Create a solved copy of instance.

Only useful for additional checks after solving.

InventoryRoutingLNS.update_instance_from_solution! — Method

update_instance_from_solution!(instance::Instance, m::Int)

Update quantities and inventories of m from the routes of instance solution.

We first update quantities received and sent from the solution and then update inventories from the quantities.

InventoryRoutingLNS.update_instance_from_solution! — Method

update_instance_from_solution!(instance::Instance)

Update quantities and inventories of all commodities from the routes of instance solution.

We first update quantities received and sent from the solution and then update inventories from the quantities.

InventoryRoutingLNS.update_instance_some_routes! — Function

update_instance_some_routes!(instance::Instance, 
                                routes::Vector{Route}, 
                                action::String, 
                                alter_solution::Bool=true
)

Locally update the instance inventories and quantities from routes.

We first apply update_quantities_some_routes! to update the quantities locally depending on the action, then propagate the quantities to inventories with update_inventories_some_routes!. When alter_solution is true we also add or delete the routes from the solution currently stored in the instance.

InventoryRoutingLNS.update_inventories_some_routes! — Method

update_inventories_some_routes!(instance::Instance, routes::Vector{Route})

Locally update the inventories from the quantities sent by routes.

We apply the same logic as in update_inventory_from_quantities! but restricted to the days, commodities, depots and customers that must be considered. The period affected is from the first departure date of the routes to the horizon T

InventoryRoutingLNS.update_inventory_from_quantities! — Method

update_inventory_from_quantities!(instance::Instance, m::Int)

Update inventories of m at depots and customers from quantities sent and received.

When the fields quantity_received and quantity_sent of instance are fixed, the daily inventory at each depot and customer can be deduced using the proper dynamics. We highlight the demand, release and routes departures are supposed to happen in the morning, whereas the reception at the customers is in the evening. Inventories are defined in the evening on each day.

InventoryRoutingLNS.update_inventory_from_quantities! — Method

update_inventory_from_quantities!(instance::Instance)

Update inventories of all commodities at depots and customers from quantities sent and received.

When the fields quantity_received and quantity_sent of instance are fixed, the daily inventory at each depot and customer can be deduced using the proper dynamics. We highlight the demand, release and routes departures are supposed to happen in the morning, whereas the reception at the customers is in the evening. Inventories are defined in the evening on each day.

InventoryRoutingLNS.update_quantities_from_solution! — Method

update_quantities_from_solution!(instance::Instance, m::Int)

Update daily quantities of commodity m sent and received from the instance solution.

Each route in the solution of instance defines deliveries that can be aggregated per customer and per depot on each day.

InventoryRoutingLNS.update_quantities_from_solution! — Method

update_quantities_from_solution!(instance::Instance)

Update daily quantities of all commodities sent and received from instance solution.

Each route in the solution of instance defines deliveries that can be aggregated per customer and per depot on each day.

InventoryRoutingLNS.update_quantities_some_routes! — Method

update_quantities_some_routes!(instance::Instance, 
                                routes::Vector{Route}, 
                                action::String
)

Locally update the quantities sent and received from routes.

We apply the same logic as in update_quantities_from_solution! but restricted to the depots and customers that are involved in routes. Depending on action, the quantities induced by routes can be added or removed.