Graphs and flows

FlowGraphs

InventoryRoutingLNS.FGN — Type

FGN

A flow graph node represents a wide array of concepts to define flow graphs.

This structure is a central block of the FlowGraph used to build commodity and vehicle flow graphs in commodity_flow_graph and expanded_vehicle_flow_graph for instance.

Fields

t::Int: date, by default 0.

t_h::Int: hour (not used in current version), by default 0.

d::Int: depot index, by default 0.

c::Int: customer index, by default 0.

s::Int: stop index, by default 0.

str::String: string to mention a concept, by default "?".

InventoryRoutingLNS.FlowGraph — Type

FlowGraph{L}

Flow graph storage with labeled vertices and ordered edges that have capacities and costs.

This structure is used to build commodity and vehicle flow graphs in commodity_flow_graph and expanded_vehicle_flow_graph for instance. It is based on the Graph.jl package.

Fields

vertexlabels::Vector{L}: vector of the vertex labels.

vertexindices::Dict{L,Int}: correspondance between vertex labels and indices.

inneighbors::Vector{Vector{Int}}: one vector of inneighbors indices per vertex index.

outneighbors::Vector{Vector{Int}}: one vector of outneighbors indices per vertex index.

edges::Vector{Graphs.SimpleEdge{Int}}: vector of edges.

edgeindices::Dict{Graphs.SimpleEdge{Int},Int}: correspondance between edges and edgeindices.

capa_min::Vector{Float64}: minimum capacity vector (on edges).

capa_max::Vector{Float64}: maximum capacity vector (on edges).

cost::Vector{Float64}: cost vector (on edges).

InventoryRoutingLNS.add_flow_constraints! — Method

add_flow_constraints!(model::Model, flowvar, g::FlowGraph)

Add flow constraints to variable flowvar of a JuMP model, based on flow graph g.

The flow constraints are the following:

minimum and maximum capacity constraints on all the arcs.

Kirchhoff constraint for the flow conservation.

InventoryRoutingLNS.add_flow_cost! — Method

add_flow_cost!(expr, flowvar, g::FlowGraph)

Add flow cost of variable flowvar to JuMP expression expr, based on flow graph g.

Commodity flow graphs

InventoryRoutingLNS.add_artificial_nodes_CFG! — Method

add_artificial_nodes_CFG!(fg::FlowGraph; refill_neighborhood::Bool)

Add the artificial nodes to the commodity flow graph fg.

These artificial nodes are used to build a flow problem.

The source and sink nodes enable a circulation.

production is a way to model and force the value of depots release with capacitated arcs.

demand does the same for the customers demand.

initial_inventory enables to set the initial inventory at each depot and each customer.

shortage_compensation is a way to model the shortage costs in the flow.

final_inventory gathers the inventory at each site in the end of the horizon T.

other_received_quantities and other_sent_quantities are only created when the boolean refill_neighborhood is true. This is done when defining a large neighborhood MILP with fill_fixed_routes_MILP.

InventoryRoutingLNS.add_customer_inventory_arcs_CFG! — Method

add_customer_inventory_arcs_CFG!(fg::FlowGraph;
                                    instance::Instance,
                                    commodity_index::Int,
                                    list_customers::AbstractVector{Int},
                                    S_max::Int,
)

Add the arcs corresponding to the customers inventory dynamics.

Thanks to the free_night nodes we manage to model the excess inventory costs. We create on each day:

one arc from morning to evening without cost or capacity.

one arc from evening to free_night with cost or capacity.

one arc from evening to morning of next day with unit excess_inventory_cost and no capacity.

one arc from free_night to morning of next day with capacity maximum_inventory.

one arc from each noon node to evening to connect the possible positions in new routes with the inventory dynamics.

InventoryRoutingLNS.add_cycle_arcs_CFG! — Method

add_cycle_arcs_CFG!(fg::FlowGraph, refill_neighborhood::Bool)

Add the arcs between artificial nodes to create the circulation.

InventoryRoutingLNS.add_depots_inventory_arcs_CFG! — Method

add_depots_inventory_arcs_CFG!(fg::FlowGraph;
                                instance::Instance,
                                commodity_index::Int,
                                list_depots::AbstractVector{Int},
)

Add the arcs corresponding to the depots inventory dynamics.

Thanks to the free_night nodes we manage to model the excess inventory costs. We create on each day:

one arc from morning to evening without cost or capacity.

one arc from evening to free_night with cost or capacity.

one arc from evening to morning of next day with unit excess_inventory_cost and no capacity.

one arc from free_night to morning of next day with capacity maximum_inventory.

InventoryRoutingLNS.add_every_day_nodes_CFG! — Method

add_every_day_nodes_CFG!(fg::FlowGraph;
                            instance::Instance,
                            list_depots::AbstractVector{Int},
                            list_customers::AbstractVector{Int},
                            S_max::Int,
)

Add the daily inventory nodes to the commodity flow graph.

We update fg with new nodes:

One morning and evening node on each day t for each site. We remind that demand, release and routes departures occur in the morning, whereas routes arrival in the evening.

One free_night node on each day to incorporate the excess inventory costs.

One noon node for each position in a new route for the customers. Most of the time S_max is 1(in this function, not in instance data) and only one position is considered: in this case we plan to build new direct routes.

InventoryRoutingLNS.add_fixed_routes_arcs_CFG! — Method

add_fixed_routes_arcs_CFG!(fg::FlowGraph;
                            instance::Instance,
                            fixed_routes::Vector{Route},
                            commodity_index::Int,
                            force_routes_values::Bool,
)

Add the arcs corresponding to the fixed_routes paths.

This function is used to integrate some routes with pre-defined paths in the flow solution. We add the nodes and arcs corresponding to those paths, and connect them to the starting depots and to the evening of the customers visited. It is particularly useful in fill_fixed_routes_MILP.

InventoryRoutingLNS.add_initial_final_inventory_arcs_CFG! — Method

add_initial_final_inventory_arcs_CFG!(fg::FlowGraph;
                                        instance::Instance,
                                        commodity_index::Int,
                                        list_depots::AbstractVector{Int},
                                        list_customers::AbstractVector{Int},
)

Add arcs between the artificial nodes and initial and final inventory nodes.

The arcs from initial_inventory to the morning of the first day are used to set the initial inventories to the values stored in instance. The final inventories are aggregated for clarity and could be used to add final inventory costs for instance.

InventoryRoutingLNS.add_new_routes_arcs_CFG! — Method

add_new_routes_arcs_CFG!(fg::FlowGraph;
                            instance::Instance,
                            list_depots::AbstractVector{Int},
                            list_customers::AbstractVector{Int},
                            commodity_index::Int,
                            maxdist::Real,
                            relaxed_trip_cost::Bool,
                            average_content_sizes::Union{Nothing,AverageContentSizes},
                            S_max::Int,
                            force_routes_values::Bool,
                            sent_quantities_to_force::Array,
                            set_upper_capa::Bool,
                            possible_delays::Dict,
                            add_delayed_arcs::Bool,
)

Add the arcs corresponding to new routes.

The maxdist argument is used to sparsify the graph, removing arcs corresponding to long distances.

When the boolean relaxed_trip_cost is true, we add costs on the routing arcs of the commodity flow graphs. This is done in a heurisic way in initialization_plus_ls! with vehicle fraction costs, and based on the statistics (saved in average_content_sizes) of a first solution in a second pass in modified_capa_initialization_plus_ls!. When used in one_step_ruin_recreate_commodity! or fill_fixed_routes_MILP, no cost is set on those routing arcs of the commodity flow graphs.

We enable forcing the quantities to send through the routes with the boolean force_routes_values and quantities sent_quantities_to_force. This is done to initialize the flow solution with the value deduced from the current IRP solution, to speed-up computations. This is done in one_step_ruin_recreate_commodity!.

For some applications in Machine Learning, we need to define a flow on a polytope (bounded) and thus to derive some upper capacities on the routing arcs. This is done when set_upper_capa is true.

Last, to derive a flow relaxation of the IRP, we need to add delayed arcs in the commodity flow graphs, which is done when add_delayed_arcs is true. In this case the direct delayed arcs have precomputed delays saved in possible_delays. The implementation of the relaxation is done in lower_bound.

InventoryRoutingLNS.add_production_arcs_depots_CFG! — Method

add_production_arcs_depots_CFG!(fg::FlowGraph;
                                instance::Instance,
                                commodity_index::Int,
                                list_depots::AbstractVector{Int},
                                refill_neighborhood::Bool,
)

Add the arcs corresponding to release and quantities to send to other customers.

We use the arcs from production to the morning of each day to set the release of each customer in accordance with the data stored in instance. Besides, since we possibly fix a part of the solution when refill_neighborhood is true, the arcs to other_sent_quantities nodes are used to take the other deliveries into account in the inventory dynamics of the depots.

InventoryRoutingLNS.add_shortage_demand_arcs_customer_CFG! — Method

add_shortage_demand_arcs_customer_CFG!(fg::FlowGraph;
                                        instance::Instance,
                                        commodity_index::Int,
                                        list_customers::AbstractVector{Int},
                                        refill_neighborhood::Bool,
)

Add the arcs corresponding to demand, shortage and quantities received.

As for the release, the arcs from each morning to the node demand are used to fix the demand in accordance with the data of instance. Besides, we create arcs from the node shortage_compensation to each morning node with shortage cost to model the soft minimum inventory constraint. Last, when refill_neighborhood is true, we also create arcs from other_received_quantities to morning nodes to take the rest of the current solution into account.

InventoryRoutingLNS.commodity_flow_graph — Method

commodity_flow_graph(instance::Instance;
                        commodity_index::Int,
                        add_new_routes_arcs::Bool = true,
                        S_max::Int = 1,
                        maxdist::Real = Inf,
                        relaxed_trip_cost::Bool = true,
                        average_content_sizes::Union{Nothing,AverageContentSizes} = nothing,
                        force_routes_values::Bool = false,
                        sent_quantities_to_force::Array = zeros(1),
                        sparsify::Bool = true,
                        set_upper_capa::Bool = false,
                        possible_delays::Dict = Dict(),
                        add_delayed_arcs::Bool = false,
                        fixed_routes::Union{Nothing, Vector{Route}} = nothing,
                        refill_neighborhood::Bool = false,
)

Create the commodity flow graph corresponding to commodity_index.

When add_new_routes_arcs is true, we add the arcs corresponding to new routes with add_new_routes_arcs_CFG!.

The argument S_max is used to restrict the length of the new routes to create. It is currently systematically set to 1. See initialization_plus_ls! and one_step_ruin_recreate_commodity! for instance.

The maxdist argument is used to sparsify the graph, removing arcs corresponding to long distances.

When sparsify is true, we restrict the nodes to the customers and depots that are concerned by the commodity with index commodity_index. We emphasize in this case the indices of the nodes and arcs differ from one commodity flow graph to the other, which makes it difficult to combine them for instance in fill_fixed_routes_MILP. We may thus disable this sparsification and force zero values in this case.

When the boolean relaxed_trip_cost is true, we add costs on the routing arcs of the commodity flow graphs. This is done in a heurisic way in initialization_plus_ls! with vehicle fraction costs, and based on the statistics (saved in average_content_sizes) of a first solution in a second pass in modified_capa_initialization_plus_ls!. When used in one_step_ruin_recreate_commodity! or fill_fixed_routes_MILP, no cost is set on those routing arcs of the commodity flow graphs.

We enable forcing the quantities to send through the routes with the boolean force_routes_values and quantities sent_quantities_to_force. This is done to initialize the flow solution with the value deduced from the current IRP solution, to speed-up computations. This is done in one_step_ruin_recreate_commodity!.

For some applications in Machine Learning, we need to define a flow on a polytope (bounded) and thus to derive some upper capacities on the routing arcs. This is done when set_upper_capa is true.

To derive a flow relaxation of the IRP, we need to add delayed arcs in the commodity flow graphs, which is done when add_delayed_arcs is true. In this case the direct delayed arcs have precomputed delays saved in possible_delays. The implementation of the relaxation is done in lower_bound.

Last, as stated before, this graph structure can be used and combined in the fill_fixed_routes_MILP function, possibly to derive a large neighborhood when the boolean refill_neighborhood is true. In this case we provide the fixed_routes to add in add_fixed_routes_arcs_CFG!.

InventoryRoutingLNS.add_artificial_nodes_CFGC! — Method

add_artificial_nodes_CFGC!(fg::FlowGraph)

Add the artificial nodes to the commodity flow graph fg of a customer.

These artificial nodes are used to build a flow problem.

The source and sink nodes enable a circulation.

production is a way to model and force the value of depots release with capacitated arcs.

demand does the same for the customers demand.

initial_inventory enables to set the initial inventory at each depot and each customer.

shortage_compensation is a way to model the shortage costs in the flow.

delivery_other_customers enables to fix the quantities to be sent by depots to other customers than the one being inserted.

final_inventory gathers the inventory at each site in the end of the horizon T.

InventoryRoutingLNS.add_customer_inventory_arcs_CFGC! — Method

add_customer_inventory_arcs_CFGC!(fg::FlowGraph;
                                    instance::Instance,
                                    customer_index::Int,
                                    commodity_index::Int,
)

Add the arcs corresponding to the customer customer_index inventory dynamics.

Thanks to the free_night nodes we manage to model the excess inventory costs. We create on each day:

one arc from morning to evening without cost or capacity.

one arc from evening to free_night with cost or capacity.

one arc from evening to morning of next day with unit excess_inventory_cost and no capacity.

one arc from free_night to morning of next day with capacity maximum_inventory.

InventoryRoutingLNS.add_cycle_arcs_CFGC! — Method

add_cycle_arcs_CFGC!(fg::FlowGraph)

Add the arcs between artificial nodes to create the circulation.

InventoryRoutingLNS.add_depots_inventory_arcs_CFGC! — Method

add_depots_inventory_arcs_CFGC!(fg::FlowGraph;
                                instance::Instance,
                                commodity_index::Int,
)

Add the arcs corresponding to the depots inventory dynamics.

Thanks to the free_night nodes we manage to model the excess inventory costs. We create on each day:

one arc from morning to evening without cost or capacity.

one arc from evening to free_night with cost or capacity.

one arc from evening to morning of next day with unit excess_inventory_cost and no capacity.

one arc from free_night to morning of next day with capacity maximum_inventory.

InventoryRoutingLNS.add_every_day_nodes_CFGC! — Method

add_every_day_nodes_CFGC!(fg::FlowGraph;
                            instance::Instance,
                            customer_index::Int,
                            routes::Vector{Route},
                            dates::Vector,
                            costs::Vector,    
)

Add the daily inventory nodes to the commodity flow graph of a customer.

We update fg with new nodes:

One morning and evening node on each day t for each site. We remind that demand, release and routes departures occur in the morning, whereas routes arrival in the evening.

One free_night node on each day to incorporate the excess inventory costs.

One node per possible position of the customer to insert in the routes. This means saving the arrival dates for each position in dates.

InventoryRoutingLNS.add_initial_final_inventory_arcs_CFGC! — Method

add_initial_final_inventory_arcs_CFGC!(fg::FlowGraph;
                                        instance::Instance,
                                        customer_index::Int,
                                        commodity_index::Int,
)

Add arcs between the artificial nodes and initial and final inventory nodes.

The arcs from initial_inventory to the morning of the first day are used to set the initial inventories to the values stored in instance. The final inventories are aggregated for clarity and could be used to add final inventory costs for instance.

InventoryRoutingLNS.add_production_delivery_arcs_depots_CFGC! — Method

add_production_delivery_arcs_depots_CFGC!(fg::FlowGraph;
                                            instance::Instance,
                                            commodity_index::Int,
)

Add the arcs corresponding to release and quantities to send to other customers.

We use the arcs from production to the morning of each day to set the release of each customer in accordance with the data stored in instance. Besides, since we fix a part of the solution when optimizing the reinsertion of one customer, the arcs to delivery_other_customers nodes are used to take the other deliveries into account in the inventory dynamics of the depots.

InventoryRoutingLNS.add_routes_arcs_CFGC! — Method

add_routes_arcs_CFGC!(fg::FlowGraph;
                        instance::Instance,
                        customer_index::Int,
                        commodity_index::Int, 
                        routes::Vector{Route},
                        dates::Vector,
                        costs::Vector,
                        force_values::Bool,
                        former_quantities_former_routes::Dict, 
                        former_quantities_new_routes::Array,
)

Add the arcs corresponding to the routes paths.

Two types of routes are represented:

new direct ones from depots to the customer with customer_index.

old routes with each possible position of the customer for reinsertion. The (routing + other customers inventories and shortage) costs induced by the reinsertion at each position are precomputed and stored in costs. The dates of arrival at each reinsertion position are also precomputed and stored in dates.

We enable forcing the quantities to send through the routes with the boolean force_values and quantities former_quantities_former_routes and former_quantities_new_routes. This is done to initialize the flow solution with the value deduced from the current IRP solution, to speed-up computations. This is done in one_step_ruin_recreate_customer!.

InventoryRoutingLNS.add_shortage_demand_arcs_customer_CFGC! — Method

add_shortage_demand_arcs_customer_CFGC!(fg::FlowGraph;
                                        instance::Instance,
                                        customer_index::Int,
                                        commodity_index::Int,
)

Add the arcs corresponding to demand and shortage.

As for the release, the arcs from each morning to the node demand are used to fix the demand in accordance with the data of instance. Besides, we create arcs from the node shortage_compensation to each morning node with shortage cost to model the soft minimum inventory constraint.

InventoryRoutingLNS.commodity_flow_graph_customer — Method

commodity_flow_graph_customer(instance::Instance;
                                customer_index::Int,
                                commodity_index::Int,
                                routes::Vector{Route},
                                dates::Vector,
                                costs::Vector,
                                force_values::Bool = false,
                                former_quantities_former_routes::Dict = Dict(), 
                                former_quantities_new_routes::Array = zeros(1),
)

Create the commodity flow graph corresponding to commodity_index for the reinsertion of the customer with index customer_index.

The (routing + other customers inventories and shortage) costs induced by the reinsertion at each position are precomputed and stored in costs. The dates of arrival at each reinsertion position are also precomputed and stored in dates. We enable forcing the quantities to send through the routes with the boolean force_values and quantities former_quantities_former_routes and former_quantities_new_routes. This is done to initialize the flow solution with the value deduced from the current IRP solution, to speed-up computations. This is done in one_step_ruin_recreate_customer!. The flowgraph is based on the structure FlowGraph.

Vehicle flow graphs

InventoryRoutingLNS.expanded_vehicle_flow_graph — Method

expanded_vehicle_flow_graph(instance::Instance; 
                                add_new_routes_arcs::Bool = true,
                                S_max::Int, 
                                maxdist::Real = Inf,
                                fixed_routes::Union{Nothing, Vector{Route}} = nothing,
                                fixed_routes_costs::Union{Nothing, Vector{Int}} = nothing,
)

Create a vehicles' flow graph with old fixed and new routes.

This function is called to derive the commodity reinsertion MILP in commodity_insertion_MILP as well as the MILP to refill fixed routes in fill_fixed_routes_MILP.

Two types of routes are represented:

new direct ones from depots to customers, added when add_new_routes_arcs is true. When creating new routes, the argument maxdist may be used to sparsify the graph by removing arcs with long distances. The maximum length of new routes is defined by S_max, which is possibly distinct from the value of instance.S_max.

old fixed_routes that visit several stops, with corresponding costs fixed_routes_costs.

InventoryRoutingLNS.expanded_vehicle_flow_graph_customer — Method

expanded_vehicle_flow_graph_customer(instance::Instance;
                                        customer_index::Int,
                                        routes::Vector{Route},
                                        cost_per_route::Vector,
                                        dates::Vector,
)

Create the vehicles' flow graph of a customer indexed by customer_index.

This function is called to derive the customer reinsertion MILP in customer_insertion_flow.

Two types of routes are represented:

new direct ones from depots to the customer with customer_index.

old routes with each possible position of the customer for reinsertion. The (routing + other customers inventories and shortage) costs induced by the reinsertion at each position are precomputed and stored in cost_per_route. The dates of arrival at each reinsertion position are also precomputed and stored in dates.