Heuristics

Bin packing

InventoryRoutingLNS.bin_packing_milp — Method

bin_packing_milp(items::Vector{IT}, weights::Vector{<:Real}, W::Real) where {IT}

Apply exact bin-packing to (items, weights) within a container of capacity W.

Solve a MILP.

InventoryRoutingLNS.first_fit_decreasing — Method

first_fit_decreasing(items::Vector{IT},
                        weights::Vector{<:Real},
                        W::Real,
) where {IT}

Apply FFD bin-packing to (items, weights) within a container of capacity W.

Delays from instance

InventoryRoutingLNS.compute_delays — Method

compute_delays(instance::Instance)

Given instance, we compute the possible delays for each depot-customer couple.

The possible delays are induced by indirect routes. This information is used to build commodity graphs with commodity_flow_graph to compute a relaxation in lower_bound.

Greedy heuristic

InventoryRoutingLNS.fill_trucks! — Method

fill_trucks!(instance::Instance; sent_quantities::Array{Int,4})

Update instance with new routes deduced from a bin packing based on sent_quantities.

InventoryRoutingLNS.initialization_plus_ls! — Method

initialization_plus_ls!(instance::Instance;
                    maxdist::Real = Inf,
                    average_content_sizes::Union{Nothing,AverageContentSizes} = nothing,
                    verbose::Bool = true,
                    reset_after::Bool = false,
                    stats::Union{Nothing, Dict} = nothing,
                    sent_quantities::Union{Nothing, Array{Int,4}} = nothing,
)

Apply a flow initialization followed by a single-depot localsearch to instance.

We first choose the quantities to send with the flow problem, initialize the IRP solution with iterative bin packin problems, and then improve the solution with single-depot moves.

The maxdist argument is used to sparsify graphs based on distances, see commodity_flow_graph for the details.

The average_content_sizes data is used when using statistics of a former pass to re-estimate the routing arc costs on the commodity flow graphs. See modified_capa_initialization_plus_ls! for the details of this process.

The reset_after boolean is to decide to reset the instance solution after this greedy_heuristic or not.

Instead of using the flows to initialize the quantities to send, we can use the sent_quantities values. This is done to show that the flows decoded from the LNS! solution are good solutions to the IRP when decoded by this function.

InventoryRoutingLNS.modified_capa_initialization_plus_ls! — Method

modified_capa_initialization_plus_ls!(instance::Instance;
                                maxdist::Real = Inf,
                                verbose::Bool = true,
                                stats::Dict = nothing,
)

Apply initialization_plus_ls! two times and update costs from statistics of the first pass.

The maxdist argument is used to sparsify graphs based on distances, see commodity_flow_graph for the details.

Relaxation

InventoryRoutingLNS.lower_bound — Method

lower_bound(instance::Instance)

Define and solve the initial flow relaxation problem on instance.

This relaxation implies the computation of possible delays to create direct delayed arcs in the commodity flow graphs. See compute_delays for the delay computation, and commodity_flow_graphfor the impact on the commodity graphs.

InventoryRoutingLNS.solve_flows_initial_solution — Method

solve_flows_initial_solution(instance::Instance; average_content_sizes::Union{Nothing,AverageContentSizes} = nothing)

Define and solve the initial flow problems to decide the quantities to send.

From the quantities we solve bin packing problems and apply localsearch in initialization_plus_ls!.

Refill routes with fixed path

InventoryRoutingLNS.decode_MILP_solution! — Method

decode_MILP_solution!(instance::Instance;
                        fixed_routes::Vector{Route},
                        flow_commodities::Matrix{Int},
                        fgs_commodities::Vector{FlowGraph},
)

Given the solution of the MILP in fill_fixed_routes_MILP, fill fixed_routes.

We use fgs_commodities to link the flow_commodities variables indices with the corresponding routes to fill in fixed_routes.

InventoryRoutingLNS.fill_fixed_routes_MILP — Method

fill_fixed_routes_MILP(instance::Instance;
                        fixed_routes::Vector{Route},
                        fixed_routes_costs::Vector{Int},
                        integer::Bool = true,
                        verbose::Bool = true,
                        refill_neighborhood::Bool = false,
                        force_quantities::Bool = false,
                        use_warm_start::Bool = false,
                        values_commodity_flows::Matrix{Int}
)

Given a set of fixed_routes, use them or not, and fill the used-ones with commodities.

When solved to optimality, this MILP is the exact decision problem of optimizing the use and deliveries inherent in a set of routes (paths). The boolean force_quantities is used to solve a degenerate MILP with sent quantities set to the previous values. It is used to provide a first solution as values_commodity_flows to a second call of this function with use_warm_start set to true, to solve the MILP with a warm start.

InventoryRoutingLNS.refill_every_route! — Method

refill_every_route!(instance::Instance; verbose::Bool = false)

Refill all the routes of a solution at once.

In this case we use refill_routes! on the whole set of routes in the current solution of instance.

We emphasize that this exact MILP may be too large to solve in practice for big IRP instances. We thus suggest some ways to fix a part of the solution and use it as large neighborhood.

InventoryRoutingLNS.refill_iterative_commodity_reinsertion! — Method

refill_iterative_commodity_reinsertion!(instance::Instance; 
                                        stats::Union{Nothing, Dict} = nothing, 
                                        verbose::Bool = false
)

Apply one_step_ruin_recreate_commodity! commodity-by-commodity.

Contrary to refill_every_route!, this iterative way of refilling every route of a solution can be used on large instances.

InventoryRoutingLNS.refill_iterative_days! — Method

refill_iterative_days!(instance::Instance; 
                        verbose::Bool = false, 
                        stats::Union{Dict, Nothing} = nothing
)

Apply refill_routes_on_day! day-by-day.

Contrary to refill_every_route!, this iterative way of refilling every route of a solution can be used on large instances.

InventoryRoutingLNS.refill_iterative_depot! — Method

refill_iterative_depot!(instance::Instance; 
                        verbose::Bool = false, 
                        stats::Union{Dict, Nothing} = nothing
)

Apply refill_routes_from_depot! depot-by-depot.

Contrary to refill_every_route!, this iterative way of refilling every route of a solution can be used on large instances.

InventoryRoutingLNS.refill_routes! — Method

refill_routes!(instance::Instance; 
                    fixed_routes::Vector{Route}, 
                    fixed_routes_costs::Vector{Int}, 
                    verbose::Bool = false,
                    refill_neighborhood::Bool = false,
                    stats::Union{Dict, Nothing} = nothing,
)

Given a set of routes, empty them, solve the refill MILP and decode the solution.

The routes to consider and corresponding costs are stored in fixed_routes and fixed_routes_costs respectively. It can be used in a large neighborhood, which entails fixing a part of the solution and creating special nodes and arcs in the commodity flow graphs (see commodity_flow_graph), indicated by the boolean refill_neighborhood.

InventoryRoutingLNS.refill_routes_from_depot! — Method

refill_routes_from_depot!(instance::Instance; 
                            depot_index::Int, 
                            verbose::Bool = false, 
                            stats::Union{Dict, Nothing} = nothing
)

Refill all the routes that start from a given depot indexed by depot_index.

In this case we use refill_routes! as a large neighborhood.

InventoryRoutingLNS.refill_routes_on_day! — Method

refill_routes_on_day!(instance::Instance; 
                        t::Int, 
                        verbose::Bool = false, 
                        stats::Union{Dict, Nothing} = nothing
)

Refill all the routes that start on a given day t.

In this case we use refill_routes! as a large neighborhood.

Matheuristic

InventoryRoutingLNS.paper_matheuristic! — Method

paper_matheuristic!(instance::Instance;
                n_it_commodity_reinsertion::Int,
                n_it_customer_reinsertion::Int, 
                tol::Float64 = 0.01,
                time_limit::Float64 = 90.0,
                verbose::Bool = false,
)

Solve the instance instance with a matheuristic;

The matheuristic consists in:

two passes of the greedy heuristic, see modified_capa_initialization_plus_ls!.

one Large Neighborhood Search defined in LNS!.

The number of commodity reinsertion per LNS iteration is n_it_commodity_reinsertion, same for n_it_customer_reinsertion and customer reinsertion. We can set a time limit to the solution process with time_limit.

Benchmark

InventoryRoutingLNS.route_based_matheuristic! — Method

route_based_matheuristic!(instance::Instance;
                        time_limit::Float64,
                        verbose::Bool = true, 
)

Adapt the ideas of recent papers such as A matheuristic algorithm for the multi-depot inventory routing problem and A Matheuristic for the Multivehicle Inventory Routing Problem to solve the multi-attribute IRP.

The benchmark works as follows:

First create an initial solution with the modified_capa_initialization_plus_ls! (flows + localsearch)

Then iteratively solve MILPS with refill_iterative_depot! to exploit the routes of the initial solution.

In the original papers one MILP is solved on a larger set of routes and all considered at once. Here because of the size of the instances and additional structure (multi-depot, multi-commodity, routes that last several days) we adapt the ideas solving one MILP per depot (heuristic).