Input-Output

Input

InventoryRoutingLNS.decompress_dataset — Method

decompress_dataset(compressed_path::String, decompressed_path::String)

Decompress the dataset located at compressed_path and save it at decompressed_path.

InventoryRoutingLNS.read_commodities_CSV — Method

read_commodities_CSV(path::String)

Create a vector of commodities from a CSV file.

InventoryRoutingLNS.read_commodity — Method

read_commodity(row::String, dims::NamedTuple)::Commodity

Create a commodity from a row of a text file.

InventoryRoutingLNS.read_customer — Method

read_customer(row::String, dims::NamedTuple)::Customer

Create a customer from a text file.

InventoryRoutingLNS.read_customers_CSV — Method

read_customers_CSV(path::String,
                    dims::NamedTuple,
                    commodity_codes::Vector{String},
                    v_index_to_code::Vector{Int},
)

Create a vector of customers from CSV files.

In each customer folder, there is one master data file, one file on maximum stock per commodity and per day, one file for the demand per commodity and day, and one file for the initial inventory, unit excess and unit shortage cost per commodity. All of them are CSV files that can be easily converted into arrays.

InventoryRoutingLNS.read_depot — Method

read_depot(row::String, dims::NamedTuple)::Depot

Create a depot from a text file.

InventoryRoutingLNS.read_depots_CSV — Method

read_depots_CSV(path::String, dims::NamedTuple, commodity_codes::Vector{String})

Create a vector of depots from CSV files.

In each depot folder, there is one master data file, one file on maximum stock per commodity and per day, one file for the release per commodity and day, and one file for the initial inventory and unit excess inventory cost per commodity. All of them are CSV files that can be easily converted into arrays.

InventoryRoutingLNS.read_dimensions — Method

read_dimensions(row::String)::NamedTuple

Read the global data of an instance from a text file.

InventoryRoutingLNS.read_dimensions_CSV — Method

read_dimensions_CSV(path::String)::NamedTuple

Read the global data of an instance from a CSV file.

InventoryRoutingLNS.read_distances — Method

read_distances(rows::Vector{String}, dims::NamedTuple)::Matrix{Int}

Get the distances matrix from a text file.

InventoryRoutingLNS.read_distances_CSV — Method

read_distances_CSV(path::String, dims::NamedTuple, v_index_to_code::Vector{Int})

Get the distances matrix from a CSV file.

We check that the indices of the rows and columns indeed correspond to the sites indices used when reading depots and customers.

InventoryRoutingLNS.read_instance — Method

read_instance(path::String)::Instance

Create a whole instance from a text file.

This function is only used on the KIRO instances (smaller and used for tests). Real instances are read from CSV files, see read_instance_CSV.

InventoryRoutingLNS.read_instance_CSV — Method

read_instance_CSV(path_to_folder::String)::Instance

Create an instance from CSV files.

An instance folder has:

one master data file.

one file on commodities.

one file on customers indices.

one file on depots indices.

one file on distances.

one file on transport durations.

one folder per customer.

one folder per depot.

InventoryRoutingLNS.read_route — Method

read_route(row::String)::Route

Create a route from a row of a text file.

InventoryRoutingLNS.read_solution — Method

read_solution(path::String)::SimpleSolution

Create a simple solution from a text file.

InventoryRoutingLNS.read_transport_durations — Method

read_transport_durations(rows::Vector{String}, dims::NamedTuple)::Matrix{Int}

Get the transport durations matrix from a distances text file.

We assume a constant speed model here.

InventoryRoutingLNS.read_transport_durations_CSV — Method

read_transport_durations_CSV(path::String, dims::NamedTuple, v_index_to_code::Vector{Int})

Get the transport durations matrix from a CSV file.

We check that the indices of the rows and columns indeed correspond to the sites indices used when reading depots and customers.

Output

InventoryRoutingLNS.Distances — Type

Distances

Store the site-to-site distances.

Fields

dist::Matrix{Int}: distances matrix (km).

Base.string — Method

Base.string(commodity::Commodity)::String

Encode commodity in a string.

Base.string — Method

Base.string(customer::Customer)::String

Encode customer in a string.

Base.string — Method

Base.string(depot::Depot)::String

Encode depot in a string.

Base.string — Method

Base.string(distances::Distances)::String

Encode site-to-site distances in a string.

Base.string — Method

Base.string(route::Route, r::Int)::String

Encode route in a string.

Base.string — Method

Base.string(stop::RouteStop)::String

Encode route stop in a string.

InventoryRoutingLNS.write_detailed_cost — Method

write_detailed_cost(; cost_dict, path_to_folder::String, name::String)

Write a detailed cost in a JSON file.

The detailed cost is studied at various steps of paper_matheuristic!.

InventoryRoutingLNS.write_instance — Method

write_instance(instance::Instance, path::String)

Encode instance in a text file at path location.

InventoryRoutingLNS.write_lower_bounds — Method

write_lower_bounds(path_to_folder::String, rescale::Bool)

Compute and write all the lower bounds on the IRP instances contained in a folder.

We possibly rescale instances with rescale_release_demand!. The lower bounds are computed with the flow relaxation lower_bound.

InventoryRoutingLNS.write_solution — Method

write_solution(instance::Instance, path::String)

Encode the solution stored in instance in a file at path location.

InventoryRoutingLNS.write_solution — Method

write_solution(solution::Solution, path::String)

Encode solution in a file at path location.

InventoryRoutingLNS.write_stats — Method

write_stats(; stats::Dict = nothing, path_to_folder::String)

Write the stats of the solution process in a JSON file for analysis.

The stats are computed during optimization and defined in paper_matheuristic!.