Large neighborhood search

Small neighborhoods

TSP generalized

InventoryRoutingLNS.TSP_neighborhood! — Method

TSP_neighborhood!(;route::Route, 
                    instance::Instance, 
                    neighborhood_index::Int, 
                    index_1::Int, 
                    index_2::Int, 
                    stats::Dict, 
                    in_LNS::Bool
)

Apply one of the three following TSP neighborhoods to route: relocate, swap, 2-opt*.

Each neighborhood involves two indices index_1 and index_2. We use neighborhood_index to choose to apply relocate (1), swap (2) or two_opt_star (3). The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_TSP_neighborhood! — Method

iterative_TSP_neighborhood!(;route::Route, 
                            instance::Instance, 
                            neighborhood_index::Int, 
                            stats::Dict, 
                            in_LNS::Bool
)

Iteratively apply a TSP neighborhood to the stops of route untill no improvement.

We use neighborhood_index to choose to apply relocate (1), swap (2) or two_opt_star (3). The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.compress! — Method

compress!(route::Route, instance::Instance)

Merge stops that visit the same customer, all gathered at the first date of stop.

InventoryRoutingLNS.compute_optimized_route_cost — Method

compute_optimized_route_cost(route::Route, instance::Instance)

Compute the cost of an optimally compressed and reordered copy of route.

InventoryRoutingLNS.optimize_route! — Method

optimize_route!(route::Route, instance::Instance)

Compress and reorder route to optimality.

InventoryRoutingLNS.optimize_route — Method

optimize_route(route::Route, instance::Instance)

Compress and reorder a copy of route to optimality.

InventoryRoutingLNS.reorder_optimally! — Method

reorder_optimally!(route::Route, instance::Instance)

Find optimal stops' order and apply it to route.

Brutal enumeration of the possible permutations.

InventoryRoutingLNS.update_route_order! — Method

update_route_order!(route::Route, instance::Instance, perm::Vector{Int})

Update a route from a permutation perm of its stops.

InventoryRoutingLNS.delete_non_profitable_routes! — Function

delete_non_profitable_routes!(instance::Instance,
                                verbose::Bool = false;
                                stats::Dict = nothing,
                                 in_LNS::Bool = true,
)

Apply delete_non_profitable_routes! every day.

InventoryRoutingLNS.delete_non_profitable_routes! — Method

delete_non_profitable_routes!(instance::Instance, 
                                t::Int, 
                                d::Int
)

Delete routes starting on day t by depot d when the total cost decreases.

Not used in the current version.

InventoryRoutingLNS.delete_non_profitable_routes! — Method

delete_non_profitable_routes!(instance::Instance,
                                t::Int;
                                stats::Dict = nothing,
                                in_LNS::Bool = true,
)

Delete routes starting on day t when the total cost decreases.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

Single-depot SDVRP generalized

InventoryRoutingLNS.get_neighbors_from_list — Method

get_neighbors_from_list(instance::Instance,
                        c::Int,
                        cs::Vector{Int},
                        prop::Float64,
)::Vector{Int}

Select the prop proportion of the closest neighbors to a customer c among customers cs.

InventoryRoutingLNS.insert_single_depot! — Method

insert_single_depot!(route1::Route,
                        route2::Route,
                        i::Int,
                        instance::Instance;
                        stats::Dict = nothing,
                        in_LNS::Bool = true,
)

Push route1.stops[i] to route2.stops and apply compress! to avoid stop repetition.

Only applied on routes with same dates of departure and same starting depot. The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.insert_swap_single_depot_routes! — Function

insert_swap_single_depot_routes!(instance::Instance,
                                    verbose::Bool = false;
                                    stats::Dict = nothing,
                                    in_LNS::Bool = true,
)

Apply iterative_insert_single_depot! and iterative_swap_single_depot! on every day, depot config.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_insert_single_depot! — Method

iterative_insert_single_depot!(instance::Instance,
                                t::Int,
                                d::Int;
                                stats::Dict = nothing,
                                in_LNS::Bool = true,
)

Try insert_single_depot! on pairs of routes sampled on day t starting at depot d.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_swap_single_depot! — Method

iterative_swap_single_depot!(instance::Instance,
                                t::Int,
                                d::Int;
                                stats::Dict = nothing,
                                in_LNS::Bool = true,
)

Try swap_single_depot! on pairs of routes sampled on day t starting at depot d.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.swap_single_depot! — Method

swap_single_depot!(route1::Route,
                    route2::Route,
                    i::Int,
                    j::Int,
                    instance::Instance;
                    stats::Dict = nothing,
                    in_LNS::Bool = true,
)

Exchange route1.stops[i] and route2.stops[j] and apply compress! to avoid stop repetition.

Only applied on routes with same dates of departure and same starting depot. The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.concatenate_multiday — Method

concatenate_multiday(routes::Vector{Route})::Route

Concatenate routes that share the same depot, shifting them all to the earliest day.

InventoryRoutingLNS.find_best_merge_multiday — Method

find_best_merge_multiday(instance::Instance, 
                            d::Int
)

Find the best pair of routes (related to the same depot d) to merge based on cost gain.

InventoryRoutingLNS.gain_merge_multiday — Method

gain_merge_multiday(route1::Route, 
                    route2::Route, 
                    instance::Instance
)::Float64

Compute the cost gain of merging two routes.

Inf if infeasible.

InventoryRoutingLNS.iterative_merge_multiday! — Function

iterative_merge_multiday!(instance::Instance,
                            verbose::Bool = false;
                            stats::Dict = nothing,
                            in_LNS::Bool = true,
)

Apply perform_best_merge_multiday! on every depot until no improvement found.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_merge_multiday! — Method

iterative_merge_multiday!(instance::Instance,
                            d::Int;
                            stats::Dict = nothing,
                            in_LNS::Bool = true,
)

Apply perform_best_merge_multiday! on depot d until no improvement found.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.perform_best_merge_multiday! — Method

perform_best_merge_multiday!(instance::Instance,
                                d::Int;
                                stats::Dict = nothing,
                                in_LNS::Bool = true,
)

Perform the best merge over the routes related to the same depot d.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.concatenate — Method

concatenate(routes::Vector{Route})

Concatenate routes that share the same depot and departure date.

InventoryRoutingLNS.evaluate_all_merges — Method

evaluate_all_merges(instance::Instance, 
                    t::Int, 
                    d::Int
)

Evaluate all the possible merges related to d and t in the current solution.

Inf if infeasible.

InventoryRoutingLNS.gain_merge — Method

gain_merge(route1::Route, 
            route2::Route, 
            instance::Instance
)::Float64

Compute the cost gain of merging two routes having same starting depot and day.

Inf if infeasible.

InventoryRoutingLNS.iterative_merge! — Function

iterative_merge!(instance::Instance,
                    verbose::Bool = false;
                    stats::Dict = nothing,
                    in_LNS::Bool = true,
)

Apply perform_best_merge! on every depot and day until no improvement found.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_merge! — Method

iterative_merge!(instance::Instance,
                    t::Int,
                    d::Int;
                    stats::Dict = nothing,
                    in_LNS::Bool = true,
)

Apply perform_best_merge! on depot d and day t until no improvement found.

The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.perform_best_merge! — Method

perform_best_merge!(instance::Instance,
                    t::Int,
                    d::Int,
                    Δ::Matrix{Float64};
                    stats::Dict = nothing,
                    in_LNS::Bool = true,
)

Perform the best merge over the routes related to the same depot d and departure date t.

In Δ we store the gain of merge for each pair of routes, iteratively updated. The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

Multi-depot SDVRP generalized

InventoryRoutingLNS.change_day! — Method

change_day!(instance::Instance, 
            route::Route, 
            new_t::Int; 
            stats::Dict = nothing
)

Change the departure day of route to new_t.

Only possible one day in the past or in the future. Feasible if the last day of arrival is before the horizon T and if the route depot has nonnegative inventory when moving to the past. Applied if feasible and cost is reduced.

InventoryRoutingLNS.change_day_all_one_pass! — Function

change_day_all_one_pass!(instance::Instance,
                            direction::String = "ahead";
                            stats::Dict = nothing,
)

Test change_day! on every route of the solution.

The direction can either be "ahead", "back" or "random". When random, we observe the realization of a Bernoulli random variable with parameter p = 0.5 and move ahead or back.

InventoryRoutingLNS.iterative_change_day! — Function

iterative_change_day!(instance::Instance, 
                        direction::String = "ahead"; 
                        stats::Dict = nothing
)

Apply change_day_all_one_pass! until no improvement is found.

InventoryRoutingLNS.insert_multi_depot! — Method

insert_multi_depot!(instance::Instance,
                    route1::Route,
                    route2::Route,
                    i::Int;
                    stats::Dict = nothing,
)

Push route1.stops[i] to route2.stops and apply compress! to avoid stop repetition.

Only applied on routes with same dates of departure.

InventoryRoutingLNS.insert_swap_multi_depot_per_day! — Method

insert_swap_multi_depot_per_day!(instance::Instance; 
                                    stats::Dict = nothing
)

Apply iterative_insert_multi_depot! and iterative_swap_multi_depot! every day.

InventoryRoutingLNS.iterative_insert_multi_depot! — Method

iterative_insert_multi_depot!(instance::Instance, 
                                t::Int; 
                                stats::Dict = nothing
)

Try insert_multi_depot! on pairs of routes sampled on day t.

InventoryRoutingLNS.iterative_swap_multi_depot! — Method

iterative_swap_multi_depot!(instance::Instance, 
                            t::Int; 
                            stats::Dict = nothing
)

Try swap_multi_depot! on pairs of routes sampled on day t.

InventoryRoutingLNS.iterative_two_opt_star_multi_depot! — Method

iterative_two_opt_star_multi_depot!(instance::Instance, 
                                    t::Int; 
                                    stats::Dict = nothing
)

Try two_opt_star_multi_depot! on pairs of routes sampled on day t.

InventoryRoutingLNS.swap_multi_depot! — Method

swap_multi_depot!(instance::Instance,
                    route1::Route,
                    route2::Route,
                    i::Int,
                    j::Int;
                    stats::Dict = nothing,
)

Exchange route1.stops[i] and route2.stops[j] and apply compress! to avoid stop repetition.

Only applied on routes with same dates of departure.

InventoryRoutingLNS.two_opt_star_multi_depot! — Method

two_opt_star_multi_depot!(instance::Instance,
                            route1::Route,
                            route2::Route,
                            k::Int,
                            l::Int;
                            stats::Dict = nothing,
)

Exchange route1.stops[k+1:end] with route2.stops[l+1:end] and compress.

Only applied on routes with same dates of departure.

InventoryRoutingLNS.two_opt_star_multi_depot_per_day! — Method

two_opt_star_multi_depot_per_day!(instance::Instance; 
                                    stats::Dict = nothing
)

Apply iterative_two_opt_star_multi_depot! every day.

Large neighborhoods

Customer reinsertion

InventoryRoutingLNS.build_model_objective_customer_insertion_flow! — Method

build_model_objective_customer_insertion_flow!(model::JuMP.Model,
                                                fg_vehicles::FlowGraph,
                                                fgs_commodities::Vector{FlowGraph},
                                                nb_m::Int,
)

Define the objective function of the customer insertion MILP based on vehicles' and commodities' flow variables.

This cost is the sum of:

one vehicles' flow cost given the arcs' costs of the fg_vehicles graph.

one commodity cost per commodity given fgs_commodities arcs costs.

InventoryRoutingLNS.coupling_constraint_customer_flow_insertion! — Method

coupling_constraint_customer_flow_insertion!(model::JuMP.Model,
                                                instance::Instance,
                                                customer_index::Int,
                                                indices_m::Vector{Any},
                                                nb_m::Int,
                                                fgs_commodities::Vector{FlowGraph},
                                                fg_vehicles::FlowGraph,
                                                old_possible_routes::Vector{Route},
                                                places::Vector{Float64},
                                                costs::Vector,
                                                dates::Vector,
)

Add the constraints binding commodities and vehicles variables related to the reinsertion of customer customer_index.

Those bindings occur for two types of routes:

new direct routes (relaxation of the capacity constraints per vehicle at the arc level).

old routes in the solution (exact constraints given the pre-computed data).

In both cases we focus on a subset of nb_m commodities indexed by indices_m

InventoryRoutingLNS.customer_insertion_flow — Method

customer_insertion_flow(instance::Instance;
                        customer_index::Int,
                        old_possible_routes::Vector{Route},
                        costs::Vector,
                        places::Vector,
                        dates::Vector,
                        former_quantities_former_routes::Dict, 
                        former_quantities_new_routes::Array,
                        use_warm_start::Bool,
                        force_values::Bool,
                        values_vehicles_flow::Array,
                        values_commodities_flows::Array,
)

Define and solve the customer insertion MILP for the customer indexed by customer_index.

We possibly force the values on former and new routes using force_values boolean and former_quantities_former_routes and former_quantities_new_routes information, see commodity_flow_graph_customer.

We can use a warmstart if use_warm_start is true, with the values of the flows set to values_commodities_flows and values_vehicles_flow.

InventoryRoutingLNS.fill_former_routes_customer_insertion! — Method

fill_former_routes_customer_insertion!(instance::Instance,
                                        old_possible_routes::Vector{Route},
                                        flows::Array{Int},
                                        commodities_of_c::Array,
                                        fg_commodity::FlowGraph,
                                        customer_index::Int,
)

Deduce from the flow variables the former routes to modify to deliver the customer indexed by customer_index.

The routes to fill are the old_possible_routes. We use the graph structure of fg_commodity to make a link between the routes' insertion positions and the indices of flows variable. We highlight the commodity flow graphs share the same arcs, thus one reference graph is usefull.

InventoryRoutingLNS.fill_new_routes_customer_insertion! — Method

fill_new_routes_customer_insertion!(instance::Instance,
                                    fg_commodity::FlowGraph,
                                    flows::Array,
                                    indices_m::Vector,
                                    nb_m::Int,
                                    customer_index::Int,
)

Create new routes solving bin packing problems involving quantities of flows.

We use the graph structure of fg_commodity to make a link between the routes to create and the indices of flows variable. Only new direct routes are created since one customer is involved.

InventoryRoutingLNS.location_and_cost_insertion! — Method

location_and_cost_insertion!(instance::Instance, c::Int, routes::Vector{Route})

Compute the costs, dates and capacities corresponding to the insertion places of customer c in routes.

For each route we compute and save the remaining loading capacity in places, the cost induced by the reinsertion of customer c (other customers' inventory and shortage costs induced by delays, and routing costs) at each position in costs, and the dates of arrival to the corresponding reinsertion positions in dates. This information is used to build both the commodities and vehicles flow graphs, see commodity_flow_graph_customer and expanded_vehicle_flow_graph_customer.

InventoryRoutingLNS.one_step_ruin_recreate_customer! — Method

one_step_ruin_recreate_customer!(instance::Instance, customer_index::Int)

Remove and insert the customer indexed by customer_index in the current solution.

We first remove the customer with remove_customer!, then precompute costs, dates and remaining capacities with location_and_cost_insertion!, solve the reinsertion MILP with customer_insertion_flow, then fill former and new routes with fill_former_routes_customer_insertion! and fill_new_routes_customer_insertion! respectively. A warmstart is performed to gain speed, initialized at the last solution.

InventoryRoutingLNS.remove_customer! — Method

remove_customer!(instance::Instance, c::Int)

Remove the customer indexed by c from the routes of the solution.

Commodity reinsertion

InventoryRoutingLNS.commodity_insertion_MILP — Method

commodity_insertion_MILP(instance::Instance;
                            commodity_index::Int,
                            integer::Bool,
                            maxdist::Real,
                            former_quantities_former_routes::Dict,
                            former_quantities_new_routes::Array,
                            force_values::Bool,
                            values_vehicles_flow::Array,
                            values_commodities_flows::Array,
                            use_warm_start::Bool
)

Define and solve the MILP for the reinsertion of commodity indexed by commodity_index.

We choose integer variables with the boolean integer.

We possibly sparsify the graphs with maxdist, and force the values on former and new routes using force_values boolean and former_quantities_former_routes and former_quantities_new_routes information, see commodity_flow_graph.

We can use a warmstart if use_warm_start is true, with the values of the flows set to values_commodities_flows and values_vehicles_flow.

InventoryRoutingLNS.fill_former_routes_commodity_insertion! — Method

fill_former_routes_commodity_insertion!(instance::Instance,
                                        old_possible_routes_indices::Vector{Int},
                                        flow_commodity::Vector{Int},
                                        fg_commodity::FlowGraph,
                                        commodity_index::Int,
)

Fill the former routes with commodity commodity_index reading the flow variables.

The routes to fill are indexed by old_possible_routes_indices. We use the graph structure of fg_commodity to make a link between the routes and the indices of the flow_commodity Vector.

InventoryRoutingLNS.fill_new_routes_commodity_insertion! — Method

fill_new_routes_commodity_insertion!(instance::Instance,
                                        fg_commodity::FlowGraph,
                                        flow_commodity::Vector{Int},
                                        commodity_index::Int,
)

Create new routes solving bin packing problems involving the flow flow_commodity.

We use the graph structure of fg_commodity to make a link between the routes to create and the indices of the flow_commodity Vector. We restrict those new routes to be direct.

InventoryRoutingLNS.one_step_ruin_recreate_commodity! — Method

one_step_ruin_recreate_commodity!(instance::Instance;
                                    commodity_index::Int,
                                    integer::Bool,
                                    maxdist::Real,
                                    delete_empty_routes::Bool = true,
)

Remove and insert the commodity indexed by commodity_index in the current solution.

We first remove the commodity with remove_commodity!, then solve the reinsertion MILP with commodity_insertion_MILP, then fill former and new routes with fill_former_routes_commodity_insertion! and fill_new_routes_commodity_insertion! respectively. A warmstart is performed to gain speed, initialized at the last solution.

InventoryRoutingLNS.remove_commodity! — Method

remove_commodity!(instance::Instance; 
                    commodity_index::Int, 
                    delete_empty_routes::Bool = true
)

Remove the commodity indexed by commodity_index from the routes of the solution.

We update the instance solution depending on delete_empty_routes. If true, the routes with empty vehicles are deleted from the solution. Else, they are kept.

Algorithms for exploration

InventoryRoutingLNS.LNS! — Method

LNS!(instance::Instance;
        tol::Real = 0.01,
        n_it_commodity_reinsertion::Int,
        n_it_customer_reinsertion::Int,
        verbose::Bool = false,
        stats::Union{Nothing, Dict} = nothing,
)

Large Neighborhood Search applied to instance.

Each iteration of the LNS is composed of:

Multi-depot local search neighborhoods for descent.

Large neighborhoods: customer and commodity reinsertion, using iterative_ruin_recreate_customer! and iterative_ruin_recreate_commodity!, as well as refill_routes! MILP-based moves.

The solution process can be stopped when the time elapsed is greater than the time limit written in stats dictionnary.

InventoryRoutingLNS.TSP_local_search! — Method

TSP_local_search!(instance::Instance;
                    niter::Int = 1,
                    verbose::Bool = false,
                    stats::Union{Nothing, Dict} = nothing,
                    in_LNS::Bool = true,
)

Apply niter for loops on the routes of the solution of instance to apply iterative_TSP_neighborhood!.

The three types of TSP neighborhoods are tested. The in_LNS boolean is used to choose in which category the statistics are included (either in the greedy initialization heuristic if false or the neighborhood of the LNS if true).

InventoryRoutingLNS.iterative_ruin_recreate_commodity! — Method

iterative_ruin_recreate_commodity!(instance::Instance;
                                    niter::Int,
                                    verbose::Bool = false,
                                    stats::Union{Nothing, Dict} = nothing,
                                    sort_by_length::Bool = false,
                                    delete_empty_routes::Bool = true,
)

Apply one_step_ruin_recreate_commodity! niter times on commodities. Two possibilities: select commodities at random, or by decreasing size.

InventoryRoutingLNS.iterative_ruin_recreate_customer! — Method

iterative_ruin_recreate_customer!(instance::Instance;
                                    niter::Int,
                                    verbose::Bool = false,
                                    at_random::Bool = false,
                                    stats::Union{Nothing, Dict} = nothing,
)

Apply one_step_ruin_recreate_customer! niter times on customers. Two possibilities: select customers at random, or by decreasing cost order.

InventoryRoutingLNS.multi_depot_local_search! — Method

multi_depot_local_search!(instance::Instance;
                            niter::Int,
                            verbose::Bool = false,
                            stats::Union{Nothing, Dict} = nothing,
                            in_LNS::Bool = true,
)

Multi-depot localsearch based on TSP, single-depot SDVRP and multi-depot SDVRP neighborhoods.

We pass niter times on each type of neighborhood.

InventoryRoutingLNS.single_depot_local_search! — Method

single_depot_local_search!(instance::Instance;
                            maxdist::Real = Inf,
                            niter::Int = 1,
                            verbose::Bool = false,
                            stats::Union{Nothing, Dict} = nothing,
                            in_LNS::Bool = true,
)

Single-depot localsearch based on merge, delete, insert and swap, and TSP neighborhoods.

We pass niter times on each type of neighborhood.