Utility functions

InventoryRoutingLNS.HEAP_TRANSPOSITIONS — Constant

HEAP_TRANSPOSITIONS

Constant list the transpositions necessary to enumerate all permutations of 1:k.

Computed for k in 1:4 only.

InventoryRoutingLNS.heap_permutations! — Method

heap_permutations!(a)

Enumerate all permutations of a in-place and print them.

InventoryRoutingLNS.heap_transpositions — Method

heap_transpositions(k::Int)

List the transpositions necessary to enumerate all permutations of 1:k.

Lightweight implementation of Heap's algorithm.

InventoryRoutingLNS.compute_total_time — Method

compute_total_time(stats::Dict)

Compute the duration in minutes elapsed since the beginning of the solution process.

The dictionnary stats stores statistics on the run.